requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pyyaml>=6.0.1
pydantic>=2.4.2